+++ auth.php
@@ -15,7 +15,4 @@
             <?php endif; ?>
         </div>
     </main>
-    <script src="vendor/components/jquery/jquery.js"></script>
-    <script src="vendor/twbs/bootstrap/dist/js/bootstrap.js"></script>
-</body>
-</html>
+    <?php include(__DIR__ . '/resources/templates/footer.php'); ?>